TypeID type refactoring.

+++ aba-industry/industry-fetcher/src/main/java/com/aba/industry/fetch/model/BlueprintOutput.java
@@ -19,5 +19,5 @@ import lombok.Data;
 public class BlueprintOutput {
     private Double  probability;
     private Integer quantity;
-    private Long    typeID;
+    private Integer typeID;
 }
